pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.2.0"
description = ""
authors = ["kat-git-hub <humanpizzza@gmail.com>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"